requirements.txt
django
django-polymorphic
django-oauth-toolkit
django-cors-middleware
django-crispy-forms
djangorestframework
